+++ src/pools/selection-strategies/less-busy-worker-choice-strategy.ts
@@ -35,7 +35,7 @@ export class LessBusyWorkerChoiceStrategy<
     opts: WorkerChoiceStrategyOptions = DEFAULT_WORKER_CHOICE_STRATEGY_OPTIONS
   ) {
     super(pool, opts)
-    this.checkOptions(opts)
+    this.checkOptions(this.opts)
   }
 
   /** @inheritDoc */